Removed old test file
